Finished Login.
Features:
	- User can enter their phone number (number only keyboard) and if the
phone number is valid (i.e. 11 digits or more as the country code is necessary
for twilio) a text is sent to the user.
	- If credentials are bad, the user will see an alert saying so.
	- User can then enter the pin.
	- If the pin is incorrect or the time is past expiration, the user
gets an alert saying so.
	- If pin is correct, the app will redirect to the home page.
	- User data is stored on the phone
	- If the user closes and opens the app, the application checks to see
if a user has logged on before. If so, while loading a technica 'splash
screen' displays while the phone loads the home page.
	- Otherwise, the login page is displayed.